LlamaIndex/FinancialStockAnalyst LlamaIndex/requirements.txt
groq
llama-index-llms-groq
ib_insync
pandas
llama-index
jinja2
streamlit
python-dotenv